Issue-255: Created new activity for View examples.
