+++ app/src/main/res/layout/a_main.xml
@@ -24,6 +24,12 @@
             android:text="@string/fragments_examples"/>
 
         <Button
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:onClick="viewsExamples"
+            android:text="@string/views_examples"/>
+
+        <Button
             android:id="@+id/fabAnimationExample"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -36,13 +42,6 @@
             android:layout_height="wrap_content"
             android:onClick="vpWithoutFragment"
             android:text="@string/vp_without_fragment_example" />
-
-        <Button
-            android:id="@+id/countDownLabel"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="countDownLabel"
-            android:text="@string/countdown_example" />
 
         <Button
             android:id="@+id/simpleTransition"
@@ -59,13 +58,6 @@
             android:text="@string/coordinator_layout_behavior_title" />
 
         <Button
-            android:id="@+id/curvedBottomBarExample"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="curvedBottomBarExample"
-            android:text="@string/curved_bottombar_example" />
-
-        <Button
             android:id="@+id/bGmailSelecting"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -73,32 +65,11 @@
             android:text="@string/gmail_selecting_example" />
 
         <Button
-            android:id="@+id/bRaindropExample"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="raindropExample"
-            android:text="@string/raindrop_view_example" />
-
-        <Button
             android:id="@+id/bConstraintLayoutAnimation"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:onClick="constraintLayoutAnimation"
             android:text="@string/constraint_layout_animation_example" />
-
-        <Button
-            android:id="@+id/bPrefixEditText"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="prefixEditText"
-            android:text="@string/prefix_edittext_example" />
-
-        <Button
-            android:id="@+id/bBottomNavigationView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="bottomNavigationView"
-            android:text="@string/bottomnavigationview_example" />
 
         <Button
             android:id="@+id/bActivityToService"
@@ -120,12 +91,6 @@
             android:layout_height="wrap_content"
             android:onClick="screenshot"
             android:text="@string/screenshot_example" />
-
-        <Button
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:onClick="progressBarCustomization"
-            android:text="@string/progress_bar_customization_example" />
 
         <Button
             android:layout_width="match_parent"
