+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,5 @@
     <string name="app_name">Example</string>
     <string name="recycler_view_examples">RecyclerView examples</string>
     <string name="fragments_examples">Fragments examples</string>
+    <string name="views_examples">Views examples</string>
 </resources>
